backend/requirements.txt
# Django
pytz==2022.1
Django==4.0.4
django-solo==2.0.0
django-filter==21.1
django-rosetta==0.9.8
django-cors-headers==3.11.0
django-import-export==2.8.0

# Static Serve
dj-static==0.0.6

# Rest framework
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0

# Swagger
drf-yasg==1.20.0

# Linters and tests
Pillow==9.1.0

